# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THE HOMIES ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SALES
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, productId, product, status, entryDate, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: string
          description: "Auto-generated like TH001"
        imei:
          type: ["string", "null"]
        product:
          type: string
        brand:
          type: ["string", "null"]
        capacity:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
          description: "Hidden from secretary"
        sellingPrice:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
          description: "Calculated: sellingPrice - purchasePrice"
        status:
          type: string
          enum: [en_stock, chez_partenaire, vendu]
        entryDate:
          type: string
          format: date
        saleDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    ProductPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [product, entryDate]
      properties:
        imei:
          type: string
        product:
          type: string
        brand:
          type: string
        capacity:
          type: string
        color:
          type: string
        supplier:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        status:
          type: string
          enum: [en_stock, chez_partenaire, vendu]
        entryDate:
          type: string
          format: date
    ProductUpdate:
      type: object
      properties:
        imei:
          type: string
        product:
          type: string
        brand:
          type: string
        capacity:
          type: string
        color:
          type: string
        supplier:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        status:
          type: string
          enum: [en_stock, chez_partenaire, vendu]
        entryDate:
          type: string

    # SALE
...